add api annotation

+++ src/main/java/com/example/TimeAttendanceAPI/service/form_record/FormRecordServiceImpl.java
@@ -138,7 +138,12 @@ public class FormRecordServiceImpl implements FormRecordService {
 
     @Override
     public void deleteFormRecord(Integer formId) {
-        formRecordRepository.deleteById(formId);
+        Optional<FormRecord> formOpt = formRecordRepository.findById(formId);
+        if (formOpt.isPresent() && formOpt.get().getStatus().equals(FormStatus.ACCEPTED)) {
+            throw new RuntimeException("Form was accepted, can't be deleted");
+        } else {
+            formRecordRepository.deleteById(formId);
+        }
     }
 
     @Override
